Add coherence metric auto-evaluation task
This adds a new Rake task to generate coherence for a given question.
Much like the answer relevancy task it:

1. generates an answer for the input question using the existing
   answer composition pipeline
2. evaluates the coherence of the generated answer against the question
   using AutoEvaluation::Coherence
3. outputs the result json to stdout
4. handles errors answers appropriately

Because there's so much shared functionality i've added a shared example
to the existing evaluation_spec to reduce duplication between the two tasks.

Once all the metrics are ported, we might want to consider updating this
so have a single rake task that takes the metric as an argument rather
than separate tasks for each metric.

I've held off on doing this for now just to make sure all the rake tasks
do have shared logic with the exception of the metric called. I'm pretty
sure they will though.