Integrate performance_analysis_consent into API endpoints
This adds the optional performance_analysis_consent boolean param to
the create and update conversation API endpoints. I've added additional
tests to make it clear that it's an optional param.

This is passed to the CreateQuestion form when present. It's defaulted
to false if not provided as the field isn't nullable in the database.

I've also updated the OpenAPI specification to document this new optional
parameter.