Move analysis-related code to answer_analysis namespace
We moved the TopicTagger to the auto_evaluation namespace in a previous commit.
This commit moves the topics job and model to an answer_analysis namespace
to better reflect its purpose.

I've run into an issue with the BigQuery individual_export code where it's
assuming that table names can be converted to model names by singularizing
and camelizing them. This doesn't work for namespaced models like
AnswerAnalysis::AnswerTopics. To fix this, I've updated the
IndividualExport class to accept model classes directly instead of table names