Rename answer analyses to answer_analysis_topics
We're moving away from having an answer analyses model to having multiple
smaller models.

The reason for this is that we need to do 3 runs per metric. Due to this
for each metric we're going to have:

- an aggregate model
- a model that represents each run

We still a model to capture answer topics so it makes sense for us to
rename the existing answer analyses model to answer topics. This renames
the table and updates all references from "answer_analysis" to
"answer_analysis_topics" where relevant.