Add AutoEvaluation::RunMetricEvaluation
We're adding a few metrics. Each of these requires a basic rake task
that can be called with an INPUT environment variable. The task will
generate a ScoreResult for the given metric and print it as JSON.

This adds the AutoEvaluation::RunMetricEvaluation class which:

- takes a question_message and a metric class as arguments
- generates an answer using the question_message,
- calls the metric class with the question and answer to get a ScoreResult
- returns the ScoreResult

If the generated answer has an error status, it returns the answer and
the rake task handles outputting the error message to stderr and aborting.