Add schema validation to BedrockOpenAIOssInvoke
This updates the BedrockOpenAIOssInvoke class to include JSON schema validation
for the structured output received from the LLM.

It uses the first tool's schema since all of our auto-eval metrics
use a single tool. If this is to change we could consider passing the
schema in via the method parameters down the line.

While doing this I noticed that i'd not been consistent with adhering to
the schemas it some test cases so i've updated those.